fix(storefront): co-locate the self-upgrade trigger with release status (BI-D77BF495)
The CEO reported the self-upgrade page "has become really hard to use...
difficult to even find where the button is to trigger it." PR #3457 already
landed the systemic UX-budget reachability axis and a safe-increment fix
(force-open the Advanced disclosure, primary-button styling) but explicitly
deferred the "fuller fix" it called out: extracting the trigger and
co-locating it with the release status card it acts on.

This completes that: SelfUpgradeTriggerControl is a new component that owns
ALL trigger/force/abort/override state and calls the existing
triggerSelfUpgrade/forceActiveRun/abortActiveRun server actions (no
duplicated quiescence/queue/override semantics). It renders inside
OwnerReleaseCard's new `primaryAction` slot, so the "Upgrade now" button is
visible on arrival in both nav modes without any disclosure needing to be
forced open. SelfUpgradeClient keeps only read-only detail (Latest Run,
history, activity, schedule, rollback) and now defaults to the ordinary
owner-first pattern: open in Full mode, collapsed by default in Simple mode.

Shared types (LatestRun/QuiescenceActivity) and isExpectedDuringSwap moved to
apps/web/lib/self-upgrade/ so the two components describe the same server
data without redefining it. Tests migrated 1:1 from SelfUpgradeClient's
trigger-control/loading/feedback describe blocks and the swap-resilience
suite into SelfUpgradeTriggerControl's own test files; added OwnerReleaseCard
coverage for the new slot and updated page.test.tsx for the co-location and
restored Simple-mode collapse behavior.

Verified locally via the documented node_modules-junction workaround for a
source-only worktree: tsc --noEmit clean (0 errors); 30/30 vitest tests
passing for SelfUpgradeTriggerControl.test.tsx + page.test.tsx. Two files
(OwnerReleaseCard.test.tsx, SelfUpgradeClient.test.tsx) hit a pre-existing
environment gap (root clone's apps/web/node_modules lacks a `next` symlink,
breaking the report-kit barrel's next/link import — present before this
change too) and the swap-resilience jsdom suite hit a pre-existing worker-
startup timeout, confirmed via an untouched file (ChangesClient.test.tsx)
failing identically. Both are environment limitations CI's real install
resolves, not regressions from this diff. Pre-commit's own scoped typecheck
hits the same missing-`next`-binary gap (DPF_SKIP_TYPECHECK=1 used here); the
manual tsc --noEmit run above is the real evidence.

Design-Grounding-Decision: extends the merged spec
docs/superpowers/specs/2026-07-22-holistic-ux-system-and-agent-codification-design.md
section 6 and docs/platform-usability-standards.md's primary-action-reachable
axis (both landed by PR #3457) with the "fuller fix" that spec explicitly
deferred — completing, not forking, the governing design.
UX-Fit-Decision: single always-visible primary action (Upgrade now), detail
(history/ledgers/logs) stays progressively disclosed; no new user-facing
config/inputs added.
Docs-Impact-Decision: updated docs/platform-usability-standards.md to record
the completed fix alongside the safe-increment PR #3457 already documented.
Module-Size-Decision: SelfUpgradeClient.tsx shrank 1459 -> 1113 LOC;
baseline retightened via scripts/check-module-size.mjs --update. New
SelfUpgradeTriggerControl.tsx (377 LOC) is well under the 800-LOC ceiling.
Process-Spine-Decision: no new process; extends the existing UX-budget spine
from PR #3457.
