+++ scripts/module-size-baseline.txt
@@ -6,8 +6,8 @@ apps/web/components/agent/AgentCoworkerPanel.tsx	1255
 apps/web/components/agent/AgentMessageInput.tsx	1034
 apps/web/components/ea/EaCanvas.tsx	1047
 apps/web/components/inventory/TopologyGraph.tsx	1031
-apps/web/components/ops/SelfUpgradeClient.test.tsx	1226
-apps/web/components/ops/SelfUpgradeClient.tsx	1459
+apps/web/components/ops/SelfUpgradeClient.test.tsx	1069
+apps/web/components/ops/SelfUpgradeClient.tsx	1113
 apps/web/components/platform/AiOperationsMap.tsx	2881
 apps/web/components/platform/EffectivePermissionsPanel.tsx	851
 apps/web/components/platform/edge-nodes/EdgeNodesAdminClient.tsx	985
@@ -37,14 +37,12 @@ apps/web/lib/explore/feature-build-types.ts	1109
 apps/web/lib/finance/ai-provider-finance.ts	1309
 apps/web/lib/inference/ai-provider-internals.ts	1228
 apps/web/lib/integrate/build-agent-prompts.ts	1064
-apps/web/lib/integrate/build-orchestrator.ts	1751
-apps/web/lib/integrate/sandbox/build-branch.ts	870
 apps/web/lib/integrate/build-orchestrator.ts	1796
-apps/web/lib/integrate/sandbox/build-branch.ts	852
+apps/web/lib/integrate/sandbox/build-branch.ts	870
 apps/web/lib/marketing.ts	1369
 apps/web/lib/mcp-tools-backlog.test.ts	874
 apps/web/lib/mcp-tools.ts	1923
-apps/web/lib/mcp/packs/backlog-pack.ts	1320
+apps/web/lib/mcp/packs/backlog-pack.ts	1073
 apps/web/lib/mcp/packs/contribution-hive-pack.ts	858
 apps/web/lib/mcp/packs/sandbox-pack.ts	920
 apps/web/lib/navigation/portal-navigation-model.ts	940
